\name{orderable-class}
\docType{class}
\alias{orderable-class}
\alias{Date-class}

\title{Class "orderable"}
\description{A class union of classes that can be used to order
  observations, in particular the periods of a simulation.  Currently,
  this includes \code{numeric}, \code{character}, \code{logical},
  \code{POSIXt}, and \code{Date}.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\author{Jeff Enos \email{jeff@kanecap.com}}
\keyword{classes}
