\name{simDataInterface-class}
\docType{class}
\alias{simDataInterface-class}

\title{Class "simDataInterface"}
\description{ Virtual class that must be extended to provide data for the
periods of a simulation.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "simDataInterface" in the signature.
}
\author{Jeff Enos \email{jeff@kanecap.com}}
\keyword{classes}
