\name{simSummaryInterface-class}
\docType{class}
\alias{simSummaryInterface-class}
\alias{simSummaryInterfaceOrNull-class}

\title{Class "simSummaryInterface"}

\description{The summary interface allows the user to specify
information to be saved out during the simulation beyond that supported
by the result classes \code{instantData} and \code{periodData}.}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"simSummaryInterfaceOrNull"}, directly.
}
\section{Methods}{
No methods defined with class "simSummaryInterface" in the signature.
}
\author{Kyle Cambell \email{Kyle.W.Campbell@williams.edu}}
\keyword{classes}
