\name{simTradesInterface-class}
\docType{class}
\alias{simTradesInterface-class}

\title{Class "simTradesInterface"}

\description{ Virtual class that must be extended to provide the
  simulator with the set of trades to execute during each period.  The
  trades returned by the interface represent the implementation of the
  trading stategy tested in the simulation.}

\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "simTradesInterface" in the signature.
}
\author{Jeff Enos \email{jeff@kanecap.com}}
\keyword{classes}
